tests/end2end/helpers: add Node, DocumentRoot(Node), Requirement(Node), Section(Node), AddNode_Menu
