+++ tests/end2end/helpers/components/node/section.py
@@ -1,0 +1,30 @@
+from selenium.webdriver.common.by import By
+from seleniumbase import BaseCase
+
+from tests.end2end.helpers.components.node.node import Node
+from tests.end2end.helpers.constants import NODE_1
+from tests.end2end.helpers.screens.document.form_edit_section import (
+    Form_EditSection,
+)
+
+
+class Section(Node):  # pylint: disable=invalid-name
+    def __init__(self, test_case: BaseCase) -> None:
+        assert isinstance(test_case, BaseCase)
+        super().__init__(test_case)
+
+    # forms
+
+    def do_open_form_edit_section(
+        self, field_order: int = NODE_1
+    ) -> Form_EditSection:
+        self.test_case.hover_and_click(
+            hover_selector=f"(//sdoc-node)[{field_order}]",
+            click_selector=(
+                f"(//sdoc-node)[{field_order}]"
+                '//*[@data-testid="node-edit-action"]'
+            ),
+            hover_by=By.XPATH,
+            click_by=By.XPATH,
+        )
+        return Form_EditSection(self.test_case)
